pyproject.toml
[project]
name = "svgizer"
version = "0.1.0"
description = "Iterative SVG refinement using OpenAI multimodal models and search-based optimization"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.21.0",
    "google-genai>=1.68.0",
    "cairosvg>=2.7.0",
    "pillow>=10.0.0",
    "numpy>=2.2.6",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0",
    "rich>=14.3.3",
    "tqdm>=4.67.3",
    "graphviz>=0.21",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-xdist",
    "ruff",
    "pyrefly",
    "py-spy",
]

[project.scripts]
svgizer = "svgizer.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-m 'not llm'"
testpaths = ["src"]
markers = [
    "llm: marks tests that require live LLM API calls",
]

[tool.pyrefly]
project-includes = ["src/**/*.py"]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu126" }]
torchvision = [{ index = "pytorch-cu126" }]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "A",
    "C4",
    "PT",
    "RET",
    "SIM",
    "PTH",
    "PIE",
    "ARG",
    "RUF",
    "TID",
]
ignore = []

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
]
